Move properties to include folder
