+++ rviz_common/CMakeLists.txt
@@ -91,24 +91,24 @@ set(rviz_common_headers_to_moc
   src/rviz_common/new_object_dialog.hpp
   src/rviz_common/panel.hpp
   src/rviz_common/panel_dock_widget.hpp
-  src/rviz_common/properties/bool_property.hpp
-  src/rviz_common/properties/color_editor.hpp
+  include/rviz_common/properties/bool_property.hpp
+  include/rviz_common/properties/color_editor.hpp
   include/rviz_common/properties/color_property.hpp
-  src/rviz_common/properties/combo_box.hpp
-  src/rviz_common/properties/editable_combo_box.hpp
-  src/rviz_common/properties/editable_enum_property.hpp
-  src/rviz_common/properties/float_edit.hpp
-  src/rviz_common/properties/float_property.hpp
-  src/rviz_common/properties/int_property.hpp
-  src/rviz_common/properties/line_edit_with_button.hpp
+  include/rviz_common/properties/combo_box.hpp
+  include/rviz_common/properties/editable_combo_box.hpp
+  include/rviz_common/properties/editable_enum_property.hpp
+  include/rviz_common/properties/float_edit.hpp
+  include/rviz_common/properties/float_property.hpp
+  include/rviz_common/properties/int_property.hpp
+  include/rviz_common/properties/line_edit_with_button.hpp
   include/rviz_common/properties/property.hpp
   # src/rviz_common/properties/property_tree_delegate.hpp
-  src/rviz_common/properties/property_tree_model.hpp
+  include/rviz_common/properties/property_tree_model.hpp
   # src/rviz_common/properties/property_tree_with_help.hpp
-  src/rviz_common/properties/status_list.hpp
-  src/rviz_common/properties/status_property.hpp
+  include/rviz_common/properties/status_list.hpp
+  include/rviz_common/properties/status_property.hpp
   include/rviz_common/properties/string_property.hpp
-  src/rviz_common/properties/tf_frame_property.hpp
+  include/rviz_common/properties/tf_frame_property.hpp
   include/rviz_common/properties/vector_property.hpp
   src/rviz_common/render_panel.hpp
   src/rviz_common/scaled_image_widget.hpp
