+++ rviz_common/include/rviz_common/display.hpp
@@ -38,8 +38,8 @@
 
 #include "rclcpp/time.hpp"
 
-#include "src/rviz_common/properties/bool_property.hpp"
-#include "src/rviz_common/properties/status_property.hpp"
+#include "rviz_common/properties/bool_property.hpp"
+#include "rviz_common/properties/status_property.hpp"
 
 class QDockWidget;
 class QWidget;
