+++ rviz_common/src/rviz_common/display.cpp
@@ -57,8 +57,8 @@
 
 #include "rviz_common/display_context.hpp"
 #include "./panel_dock_widget.hpp"
-#include "./properties/property_tree_model.hpp"
-#include "./properties/status_list.hpp"
+#include "rviz_common/properties/property_tree_model.hpp"
+#include "rviz_common/properties/status_list.hpp"
 #include "./window_manager_interface.hpp"
 
 namespace rviz_common
