+++ rviz_common/src/rviz_common/failed_display.cpp
@@ -33,7 +33,7 @@
 #include <QColor>  // NOLINT: cpplint is unable to handle the include order here
 
 #include "./load_resource.hpp"
-#include "./properties/status_property.hpp"
+#include "rviz_common/properties/status_property.hpp"
 
 namespace rviz_common
 {
