+++ rviz_common/src/rviz_common/properties/color_property.cpp
@@ -32,9 +32,9 @@
 #include <QStringList>
 #include <QStyleOptionViewItem>
 
-#include "./parse_color.hpp"
+#include "rviz_common/properties/parse_color.hpp"
 #include "rviz_common/properties/color_property.hpp"
-#include "./color_editor.hpp"
+#include "rviz_common/properties/color_editor.hpp"
 
 namespace rviz_common
 {
