+++ rviz_common/src/rviz_common/properties/combo_box.cpp
@@ -27,7 +27,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "./combo_box.hpp"
+#include "rviz_common/properties/combo_box.hpp"
 
 namespace rviz_common
 {
