+++ rviz_common/src/rviz_common/properties/enum_property.cpp
@@ -32,7 +32,7 @@
 
 #include <string>
 
-#include "./combo_box.hpp"
+#include "rviz_common/properties/combo_box.hpp"
 
 namespace rviz_common
 {
