+++ rviz_common/src/rviz_common/properties/property.cpp
@@ -37,8 +37,8 @@
 #include <QLineEdit>  // NOLINT: cpplint is unable to handle the include order here
 #include <QSpinBox>  // NOLINT: cpplint is unable to handle the include order here
 
-#include "./float_edit.hpp"
-#include "./property_tree_model.hpp"
+#include "rviz_common/properties/float_edit.hpp"
+#include "rviz_common/properties/property_tree_model.hpp"
 
 namespace rviz_common
 {
