+++ rviz_common/src/rviz_common/temp/default_plugins/displays/grid_display.cpp
@@ -49,7 +49,7 @@
 #include "rviz_common/display_context.hpp"
 #include "rviz_common/frame_manager.hpp"
 #include "rviz_rendering/grid.hpp"
-#include "../../../properties/parse_color.hpp"
+#include "rviz_common/properties/parse_color.hpp"
 #include "rviz_common/properties/property.hpp"
 #include "../../../selection/selection_manager.hpp"
