+++ rviz_common/src/rviz_common/temp/default_plugins/displays/tf_display.cpp
@@ -56,8 +56,8 @@
 #include "rviz_rendering/arrow.hpp"
 #include "rviz_rendering/axes.hpp"
 #include "rviz_rendering/movable_text.hpp"
-#include "../../../properties/bool_property.hpp"
-#include "../../../properties/float_property.hpp"
+#include "rviz_common/properties/bool_property.hpp"
+#include "rviz_common/properties/float_property.hpp"
 #include "rviz_common/properties/quaternion_property.hpp"
 #include "rviz_common/properties/string_property.hpp"
 #include "rviz_common/properties/vector_property.hpp"
