+++ rviz_common/src/rviz_common/temp/default_plugins/view_controllers/orbit_view_controller.cpp
@@ -54,8 +54,8 @@
 
 #include "rviz_common/display_context.hpp"
 #include "../../../load_resource.hpp"
-#include "../../../properties/bool_property.hpp"
-#include "../../../properties/float_property.hpp"
+#include "rviz_common/properties/bool_property.hpp"
+#include "rviz_common/properties/float_property.hpp"
 #include "../../../render_panel.hpp"
 #include "../../../viewport_mouse_event.hpp"
 #include "rviz_common/properties/vector_property.hpp"
