+++ rviz_common/src/rviz_common/temp/robot/robot_link.cpp
@@ -67,8 +67,8 @@
 #include "rviz_rendering/axes.hpp"
 #include "rviz_rendering/object.hpp"
 #include "rviz_rendering/shape.hpp"
-#include "../../properties/float_property.hpp"
-#include "../../properties/bool_property.hpp"
+#include "rviz_common/properties/float_property.hpp"
+#include "rviz_common/properties/bool_property.hpp"
 #include "rviz_common/properties/property.hpp"
 #include "rviz_common/properties/quaternion_property.hpp"
 #include "rviz_common/properties/vector_property.hpp"
