+++ rviz_common/src/rviz_common/visualization_manager.cpp
@@ -77,12 +77,12 @@
 // #include "./ogre_helpers/qt_ogre_render_window.hpp"
 // #include "./ogre_helpers/render_system.hpp"
 #include "rviz_common/properties/color_property.hpp"
-#include "./properties/int_property.hpp"
-#include "./properties/parse_color.hpp"
+#include "rviz_common/properties/int_property.hpp"
+#include "rviz_common/properties/parse_color.hpp"
 #include "rviz_common/properties/property.hpp"
-#include "./properties/property_tree_model.hpp"
-#include "./properties/status_list.hpp"
-#include "./properties/tf_frame_property.hpp"
+#include "rviz_common/properties/property_tree_model.hpp"
+#include "rviz_common/properties/status_list.hpp"
+#include "rviz_common/properties/tf_frame_property.hpp"
 #include "./render_panel.hpp"
 #include "./selection/selection_manager.hpp"
 // #include "./tool.hpp"
